requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
google-generativeai==0.8.3
sentence-transformers==3.2.0
torch==2.4.1
transformers==4.45.2
faiss-cpu==1.9.0
pandas==2.1.4
scikit-learn==1.3.2
numpy==1.26.4
nltk==3.9.1
huggingface-hub==0.26.0
pdfplumber==0.11.4  # Better PDF text extraction
pytesseract==0.3.10  # For OCR (optional)
pillow==11.1.0  # Required for image processing in OCR